cleaning old kronmult logic
