+++ src/asgard_kronmult.hpp
@@ -7,6 +7,61 @@
 namespace asgard::kronmult
 {
 /*!
+  * \brief Compute the permutations (upper/lower) for global kronecker operations
+  *
+  * This computes all the permutations for the given dimensions
+  * and sets up the fill and direction vector-of-vectors.
+  * Direction 0 will be set to full and all others will alternate
+  * between upper and lower.
+  *
+  * By default, the directions are in order (0, 1, 2, 3); however, if a term has
+  * entries (identity, term, identity, term), then the effective dimension is 2
+  * and first the permutation should be set for dimension 2,
+  * then we should call .remap_directions({1, 3}) to remap (0, 1) into the active
+  * directions of 1 and 3 (skipping the call to the identity.
+ */
+struct permutes
+{
+  //! \brief Matrix fill for each operation.
+  std::vector<std::vector<conn_fill>> fill;
+  //! \brief Direction for each matrix operation.
+  std::vector<std::vector<int>> direction;
+  //! \brief Direction of the flux, if any
+  int flux_dir = -1;
+  //! \brief Empty permutation list.
+  permutes() = default;
+  //! \brief Initialize the permutations.
+  permutes(int num_dimensions);
+  //! \brief Create uniform transformation, only lower or upper.
+  permutes(int num_dimensions, conn_fill same_fill);
+  //! Creates a transformation with the specified active and flux directions.
+  permutes(std::vector<int> const &active_dirs, int fdir = -1)
+      : permutes(static_cast<int>(active_dirs.size()))
+  {
+    remap_directions(active_dirs);
+    flux_dir = fdir;
+  }
+  //! \brief (debugging) Convert the fill to a string.
+  std::string_view fill_name(int perm, int stage) const;
+  //! \brief Shows the number of dimensions considered in the permutation
+  int num_dimensions() const
+  {
+    return (direction.empty()) ? 0 : static_cast<int>(direction.front().size());
+  }
+  //! \brief Reindexes the dimensions to match the active (non-identity) dimensions
+  void remap_directions(std::vector<int> const &active_dirs)
+  {
+    for (auto &dirs : direction) // for all permutations
+      for (auto &d : dirs)       // for all directions
+        d = active_dirs[d];
+  }
+  //! \brief Pads all permutations with the given dimensions and assuming upper matrices
+  void prepad_upper(std::vector<int> const &additional);
+  //! \brief Indicates if the permutation has been set
+  operator bool () const { return not direction.empty(); }
+};
+
+/*!
  * \internal
  * \brief Persistent workspace for kronmult operations
  *
